Get the Emacs to language mapping list up and going
I've put this together by going though LEXER and looking for the mode that
matches. Where there isn't one I've left a comment as a reminder for someone
to fill in that blank.

I've also included a reminder for hylang.org as it might be nice to have it
in here too.